修复编译错误

+++ src/utilities.cpp
@@ -190,5 +190,5 @@ int intTransition(int from, int to, long tickStart, long duration, long tickNow)
 	if (tickNow<0) tickNow = SDL_GetTicks();
 	float elapsed = (float)(tickNow-tickStart)/duration;
 	//                    elapsed                 increments
-	return constrain(round(elapsed*(to-from)),from,to);
+	return constrain((int)round(elapsed*(to-from)),from,to);
 }
